Incorrect solution to mincut algorithm
Would work, but a case like
a --- b --- c
  \ /  \  /
   d    e

fails, because current implementaiotn sees
abc, abe, and then wehn adbc, adbec attempt to get to c,
they contain redundant edges (bc or be),
and so they are discarded.
But when abe attempts to get to c via abec,
it sees that abc has already used ab, and so this
path is ignored. But since the path through e will be
ignored when it goes to c, that means the path adbcec should
be allowed, b/c even though it goes through e, it shares
no edges in common with abc, so there are 2 ways to get from a to c,
and we must remove an edge from each of them.